Rename, delete servlets
